pyproject.toml
[tool.poetry]
name = "autototp-server"
version = "0.1.0"
description = "Auto TOTP server"
authors = ["Johnny Wezel <j@wezel.name>"]
license = "GPL2"
readme = "README.md"
packages = [{include = "autototp"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
better-exceptions = "^0.3.3"
pyotp = "^2.8.0"
uvicorn = "^0.22.0"
loguru = "^0.7.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.4.2"


[[tool.poetry.source]]
name = "default"
url = "https://pypi.org"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
